move thunk into its own folder
